requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
cx-Oracle==7.3.0
Flask==1.1.1
Flask-Cors==3.0.8
flask-restx==0.1.1
Flask-SQLAlchemy==2.4.1
gevent==1.4.0
greenlet==0.4.15
idna==2.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==8.2.0
Naked==0.1.31
packaging==20.1
pluggy==0.13.1
py==1.8.1
pycparser==2.19
pycryptodome==3.9.6
PyJWT==1.7.1
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
pytest-flask==0.15.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
shellescape==3.8.1
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==0.16.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==2.1.0
